\name{CollectData}
\alias{CollectData}
\docType{data}
\title{CollectData
}
\description{
    Including data for neural-tube birth defects (NTD) Y and suspected and environmental factor data, "elevation", "soil type", and "watershed".
}
\usage{data("CollectData")}
\format{
  A data frame with 185 observations on the following 4 variables.
}
\keyword{datasets}
