% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecological_detector.r
\name{ecological_detector}
\alias{ecological_detector}
\title{ecological detector}
\usage{
ecological_detector(y_column, x_column_nn, tabledata)
}
\arguments{
\item{y_column}{The index or field name of explained variable column in input dataset.}

\item{x_column_nn}{The index or field name of explanatory variable(s)in input dataset.}

\item{tabledata}{The dataset (dataframe) contains fields of explained variable and explanatory variables.}
}
\value{
Results of ecological detector is the significance test of impact difference between two explanatory variables.
}
\description{
This function identifies the impact differences between two factors X1 ~ X2.
}
\examples{
data(CollectData)
ecological_detector("incidence",c("type","region"),CollectData)
ecological_detector("incidence",c("type","region","level"),CollectData)
}
\keyword{detector}
\keyword{ecological}
