% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_detector.r
\name{factor_detector}
\alias{factor_detector}
\title{factor detector}
\usage{
factor_detector(y_column, x_column_nn, tabledata)
}
\arguments{
\item{y_column}{The index or field name of explained variable in input dataset.}

\item{x_column_nn}{The index or the field name(s) of explanatory variable(s) in input dataset.}

\item{tabledata}{The dataset (dataframe) contains fields of explained variable and explanatory variables.}
}
\value{
Results of factor detector include q statistic and the corresponding p value.
}
\description{
The factor detector q-statistic measures the spatial stratified heterogeneity of a variable Y, or the determinant power of a covariate X of Y.
}
\examples{
data(CollectData)
factor_detector("incidence","soiltype",CollectData)
factor_detector(1,2,CollectData)
factor_detector (1,c(2,3,4),CollectData)
factor_detector ("incidence",c("soiltype","watershed"),CollectData)
}
\keyword{detector}
\keyword{factor}
