% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_detector.r
\name{interaction_detector}
\alias{interaction_detector}
\title{interaction detector}
\usage{
interaction_detector(y_column, x_column_nn, tabledata)
}
\arguments{
\item{y_column}{The index or field name of explained variable in input dataset.}

\item{x_column_nn}{The index or field name of explanatory variable(s) in input dataset.}

\item{tabledata}{The dataset (dataframe) contains fields of explained variable and explanatory variables.}
}
\value{
Results of interaction detector include the interactive q satistic.
}
\description{
This function reveals whether the risk factors X1 and X2 (and more X) have an interactive influence on a disease Y.
}
\examples{
data(CollectData) 
interaction_detector("incidence",c("type","region"),CollectData)
interaction_detector("incidence",c("type","region","level"),CollectData)
}
\keyword{detector}
\keyword{interaction}
